Comments added, errors or bugs tested.
App is fully functional as per grading criteria without any erros or bugs.
Tested.

● Custom dice AND previous dice
rolls are saved
● Custom dice AND previous rolls
can be cleared (un-saved)
● External (Die.java) classes are
used efficiently to manage dice
rolls
Users can enter and roll custom
dice
● Previous dice rolls are displayed
to the user
The layout supports custom dice
● A historical view of previous dice
is available
● The layout is easy to use
The theme and individual styles
are combined in a way to make
the app look professional
● App takes advantage of normal
and night themes